+++ app/home/historical/page.tsx
@@ -4,8 +4,10 @@ import Pagination from "@/app/ui/pagination";
 import Search from "@/app/ui/search";
 import MetricsTable from "@/app/ui/historical/metricsTable";
 import YearRangeButtons from "@/app/ui/historical/button/index";
+import Map from "@/app/ui/historical/map";
 import Footer from "@/app/ui/footer";
 import ErrorMessage from "@/app/ui/error";
+import { formatDateToLocal } from "@/app/lib/utils";
 
 export default async function getServerSideProps(props: {
   searchParams?: Promise<{
@@ -43,7 +45,7 @@ export default async function getServerSideProps(props: {
         End Date: {endDate}
       </div>
       <Search placeholder="Search Lighthouse..." />
-      {lighthouses.map(async(lighthouse) => {
+      {lighthouses.map(async(lighthouse, index) => {
         const { daily } = await fetchHistoricalWeather(
           lighthouse.latitude, 
           lighthouse.longitude, 
@@ -57,21 +59,38 @@ export default async function getServerSideProps(props: {
         const maxGustValue = Math.max(...maxGust);
         return (
           <>
-          <div className="grid grid-cols-8 grid-rows-2 gap-1 flex items-center border-2 shadow-md">
-          <div className="col-span-1 row-span-1 text-lg sm:text-xl md:text-2xl text-center tracking-wide text-blue-600 dark:text-sky-400">
-            {lighthouse.name}
-          </div>
-
-              <div className="col-span-7 row-span-1">
-              <MetricsTable 
-                maxGust={maxGustValue} 
-                maxWind={maxWindValue}
-              />
+            <div key={lighthouse.name} className="grid grid-cols-1 md:grid-cols-8 grid-rows-3 md:grid-rows-[auto_auto] gap-4 border-2 shadow-md p-4 md:p-6">
+              <div className="col-span-1 md:col-span-8 row-span-1 text-lg md:text-3xl text-center tracking-wide text-blue-600 dark:text-sky-400">
+                {lighthouse.name}
               </div>
-              <div className="col-span-8 row-span-1">
+              <div className="col-span-1 md:col-span-3 row-span-1">
+                <Map
+                  key={index}
+                  id={lighthouse.id}
+                  name={lighthouse.name}
+                  latitude={lighthouse.latitude}
+                  longitude={lighthouse.longitude}
+                  abovewater={lighthouse.abovewater}
+                  towerheight={lighthouse.towerheight}
+                  range_w={lighthouse.range_w}
+                  range_r={lighthouse.range_r}
+                  coast={lighthouse.coast}
+                  constructed={formatDateToLocal(lighthouse.constructed)}
+                  currentdate={lighthouse.currentdate}
+                  age={lighthouse.age}
+                  image_url={lighthouse.image_url}
+                />
+              </div>
+              <div className="col-span-1 md:col-span-5 row-span-1">
+                <MetricsTable 
+                  maxGust={maxGustValue} 
+                  maxWind={maxWindValue}
+                />
+              </div>
+              <div className="col-span-1 md:col-span-8 row-span-1">
                 <LineGraph daily = {daily}/>
               </div>
-          </div>
+            </div>
           </>
         )
       })}