+++ app/ui/historical/map/index.tsx
@@ -1,0 +1,13 @@
+'use client';
+import dynamic from "next/dynamic";
+
+
+const Map = dynamic(
+    () => import('@/app/ui/historical/map/map'),
+    { 
+      loading: () => <p>A map is loading</p>,
+      ssr: false
+    }
+)
+
+export default Map;