+++ app/ui/home/metricsTable.tsx
@@ -22,36 +22,36 @@ export default function MetricsTable({
 }: LighthouseMetrcsPropr) {
 
   return (
-    <table className="min-w-full table-auto border-collapse border border-gray-200 px-4 py-2 border-b text-left">
-    <thead>
-      <tr>
-        <th>Lighthouse</th>
-        <th>Type</th>
-        <th>Max Metric</th>
-      </tr>
-    </thead>
-    <tbody>
-      <tr>
-        <td>{rangename}</td>
-        <td>Range</td>
-        <td>{maxrange} N</td>
-      </tr>
-      <tr>
-        <td>{towerheightname}</td>
-        <td>Tower Height</td>
-        <td>{maxtowerheight} Meters</td>
-      </tr>
-      <tr>
-        <td>{abovewatername}</td>
-        <td>Above Water</td>
-        <td>{maxabovewater} Meters</td>
-      </tr>
-      <tr>
-        <td>{agename}</td>
-        <td>Age</td>
-        <td>{maxage} Years</td>
-      </tr>
-    </tbody>
-  </table>
+    <table className="min-w-full table-auto border-collapse border border-darkgray px-4 py-2 border-b text-left text-maritime">
+  <thead style={{ backgroundColor: '#ADD8E6' }}>
+    <tr>
+      <th className="px-4 py-2 border-b border-darkgray">LIGHTHOUSE</th>
+      <th className="px-4 py-2 border-b border-darkgray">TYPE</th>
+      <th className="px-4 py-2 border-b border-darkgray">MAX METRIC</th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr style={{ backgroundColor: '#E0FFFF' }}>
+      <td className="px-4 py-2 border-b border-darkgray">{rangename}</td>
+      <td className="px-4 py-2 border-b border-darkgray">RANGE (W)</td>
+      <td className="px-4 py-2 border-b border-darkgray">{maxrange} NM</td>
+    </tr>
+    <tr style={{ backgroundColor: '#F0FFF0' }}>
+      <td className="px-4 py-2 border-b border-darkgray">{towerheightname}</td>
+      <td className="px-4 py-2 border-b border-darkgray">TWR HT</td>
+      <td className="px-4 py-2 border-b border-darkgray">{maxtowerheight} M</td>
+    </tr>
+    <tr style={{ backgroundColor: '#E0FFFF' }}>
+      <td className="px-4 py-2 border-b border-darkgray">{abovewatername}</td>
+      <td className="px-4 py-2 border-b border-darkgray">ABV WTR</td>
+      <td className="px-4 py-2 border-b border-darkgray">{maxabovewater} M</td>
+    </tr>
+    <tr style={{ backgroundColor: '#F0FFF0' }}>
+      <td className="px-4 py-2 border-b border-darkgray">{agename}</td>
+      <td className="px-4 py-2 border-b border-darkgray">AGE</td>
+      <td className="px-4 py-2 border-b border-darkgray">{maxage} YRS</td>
+    </tr>
+  </tbody>
+</table>
   );
 }